requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.36
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.2
httpx==0.25.2
pytest==8.3.4
urllib3==2.1.0
